Enable JSON schema validation for import
