+++ src/AppInstallerCLICore/pch.h
@@ -7,6 +7,14 @@
 #include <shellapi.h>
 #include <WinInet.h>
 
+#pragma warning( push )
+#pragma warning ( disable : 4458 4100 4702 )
+#include <valijson/schema.hpp>
+#include <valijson/schema_parser.hpp>
+#include <valijson/validator.hpp>
+#include <valijson/adapters/jsoncpp_adapter.hpp>
+#pragma warning( pop )
+
 #include <array>
 #include <iostream>
 #include <fstream>
