+++ src/AppInstallerCommonCore/Manifest/ManifestSchemaValidation.cpp
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 #include "pch.h"
 #include "winget/Yaml.h"
+#include "winget/JsonSchemaValidation.h"
 #include "winget/ManifestCommon.h"
 #include "winget/ManifestSchemaValidation.h"
 #include "winget/ManifestYamlParser.h"
@@ -86,35 +87,6 @@ namespace AppInstaller::Manifest::YamlParser
         }
     }
 
-    std::string LoadResourceAsString(PCWSTR resourceName, PCWSTR resourceType)
-    {
-        HMODULE resourceModule = NULL;
-        GetModuleHandleEx(
-            GET_MODULE_HANDLE_EX_FLAG_FROM_ADDRESS,
-            (PCWSTR)LoadResourceAsString,
-            &resourceModule);
-        THROW_LAST_ERROR_IF_NULL(resourceModule);
-
-        HRSRC resourceInfoHandle = FindResource(resourceModule, resourceName, resourceType);
-        THROW_LAST_ERROR_IF_NULL(resourceInfoHandle);
-
-        HGLOBAL resourceMemoryHandle = LoadResource(resourceModule, resourceInfoHandle);
-        THROW_LAST_ERROR_IF_NULL(resourceMemoryHandle);
-
-        ULONG resourceSize = 0;
-        char* resourceContent = NULL;
-        resourceSize = SizeofResource(resourceModule, resourceInfoHandle);
-        THROW_LAST_ERROR_IF(resourceSize == 0);
-
-        resourceContent = reinterpret_cast<char*>(LockResource(resourceMemoryHandle));
-        THROW_HR_IF_NULL(E_UNEXPECTED, resourceContent);
-
-        std::string resourceStr;
-        resourceStr.assign(resourceContent, resourceSize);
-
-        return resourceStr;
-    }
-
     Json::Value LoadSchemaDoc(const ManifestVer& manifestVersion, ManifestTypeEnum manifestType)
     {
         std::string schemaStr;
@@ -124,19 +96,19 @@ namespace AppInstaller::Manifest::YamlParser
             switch (manifestType)
             {
             case AppInstaller::Manifest::ManifestTypeEnum::Singleton:
-                schemaStr = LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_SINGLETON), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
+                schemaStr = JsonSchema::LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_SINGLETON), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
                 break;
             case AppInstaller::Manifest::ManifestTypeEnum::Version:
-                schemaStr = LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_VERSION), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
+                schemaStr = JsonSchema::LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_VERSION), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
                 break;
             case AppInstaller::Manifest::ManifestTypeEnum::Installer:
-                schemaStr = LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_INSTALLER), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
+                schemaStr = JsonSchema::LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_INSTALLER), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
                 break;
             case AppInstaller::Manifest::ManifestTypeEnum::DefaultLocale:
-                schemaStr = LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_DEFAULTLOCALE), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
+                schemaStr = JsonSchema::LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_DEFAULTLOCALE), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
                 break;
             case AppInstaller::Manifest::ManifestTypeEnum::Locale:
-                schemaStr = LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_LOCALE), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
+                schemaStr = JsonSchema::LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_V1_LOCALE), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
                 break;
             default:
                 THROW_HR(HRESULT_FROM_WIN32(ERROR_NOT_SUPPORTED));
@@ -144,19 +116,10 @@ namespace AppInstaller::Manifest::YamlParser
         }
         else
         {
-            schemaStr = LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_PREVIEW), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
+            schemaStr = JsonSchema::LoadResourceAsString(MAKEINTRESOURCE(IDX_MANIFEST_SCHEMA_PREVIEW), MAKEINTRESOURCE(MANIFESTSCHEMA_RESOURCE_TYPE));
         }
 
-        Json::Value schemaJson;
-        int schemaLength = static_cast<int>(schemaStr.length());
-        Json::CharReaderBuilder charReaderBuilder;
-        const std::unique_ptr<Json::CharReader> jsonReader(charReaderBuilder.newCharReader());
-        std::string errorMsg;
-        if (!jsonReader->parse(schemaStr.c_str(), schemaStr.c_str() + schemaLength, &schemaJson, &errorMsg)) {
-            THROW_HR_MSG(E_UNEXPECTED, "Jsoncpp parser failed to parse the schema doc. Reason: %s", errorMsg.c_str());
-        }
-
-        return schemaJson;
+        return JsonSchema::LoadSchemaDoc(schemaStr);
     }
 
     std::vector<ValidationError> ValidateAgainstSchema(const std::vector<YamlManifestInfo>& manifestList, const ManifestVer& manifestVersion)
@@ -164,7 +127,6 @@ namespace AppInstaller::Manifest::YamlParser
         std::vector<ValidationError> errors;
         // A list of schema validator to avoid multiple loadings of same schema
         std::map<ManifestTypeEnum, valijson::Schema> schemaList;
-        valijson::Validator schemaValidator;
 
         for (const auto& entry : manifestList)
         {
@@ -173,36 +135,17 @@ namespace AppInstaller::Manifest::YamlParser
                 // Copy constructor of valijson::Schema was private
                 valijson::Schema& newSchema = schemaList.emplace(
                     std::piecewise_construct, std::make_tuple(entry.ManifestType), std::make_tuple()).first->second;
-                valijson::SchemaParser schemaParser;
                 Json::Value schemaJson = LoadSchemaDoc(manifestVersion, entry.ManifestType);
-                valijson::adapters::JsonCppAdapter jsonSchemaAdapter(schemaJson);
-                schemaParser.populateSchema(jsonSchemaAdapter, newSchema);
+                JsonSchema::PopulateSchema(schemaJson, newSchema);
             }
 
             const auto& schema = schemaList.find(entry.ManifestType)->second;
-
             Json::Value manifestJson = ManifestYamlNodeToJson(entry.Root);
-            valijson::adapters::JsonCppAdapter manifestJsonAdapter(manifestJson);
             valijson::ValidationResults results;
 
-            if (!schemaValidator.validate(schema, manifestJsonAdapter, &results))
+            if (!JsonSchema::Validate(schema, manifestJson, results))
             {
-                valijson::ValidationResults::Error error;
-                std::stringstream ss;
-
-                ss << "Schema validation failed." << std::endl;
-                while (results.popError(error))
-                {
-                    std::string context;
-                    for (auto itr = error.context.begin(); itr != error.context.end(); itr++)
-                    {
-                        context += *itr;
-                    }
-
-                    ss << "Error context: " << context << " Description: " << error.description << std::endl;
-                }
-
-                errors.emplace_back(ValidationError::MessageWithFile(ss.str(), entry.FileName));
+                errors.emplace_back(ValidationError::MessageWithFile(JsonSchema::GetErrorStringFromResults(results), entry.FileName));
             }
         }
