+++ src/Schemas/PackagesSchema.h
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+#pragma once
+
+#define PACKAGESSCHEMA_RESOURCE_TYPE    300
+
+#define IDX_PACKAGES_SCHEMA_V1          301
