+++ src/Schemas/Schemas.rc
@@ -2,6 +2,7 @@
 //
 #include "resource.h"
 #include "ManifestSchema.h"
+#include "PackagesSchema.h"
 
 #define APSTUDIO_READONLY_SYMBOLS
 /////////////////////////////////////////////////////////////////////////////
@@ -69,3 +70,9 @@ IDX_MANIFEST_SCHEMA_V1_VERSION            MANIFESTSCHEMA_RESOURCE_TYPE    "..\\.
 IDX_MANIFEST_SCHEMA_V1_INSTALLER          MANIFESTSCHEMA_RESOURCE_TYPE    "..\\..\\schemas\\JSON\\manifests\\v1.0.0\\manifest.installer.1.0.0.json"
 IDX_MANIFEST_SCHEMA_V1_DEFAULTLOCALE      MANIFESTSCHEMA_RESOURCE_TYPE    "..\\..\\schemas\\JSON\\manifests\\v1.0.0\\manifest.defaultLocale.1.0.0.json"
 IDX_MANIFEST_SCHEMA_V1_LOCALE             MANIFESTSCHEMA_RESOURCE_TYPE    "..\\..\\schemas\\JSON\\manifests\\v1.0.0\\manifest.locale.1.0.0.json"
+
+/////////////////////////////////////////////////////////////////////////////
+//
+// Packages schema
+//
+IDX_PACKAGES_SCHEMA_V1                    PACKAGESSCHEMA_RESOURCE_TYPE    "..\\..\\schemas\\JSON\\packages\\packages.schema.1.0.json"
