+++ src/Schemas/resource.h
@@ -1,6 +1,6 @@
 //{{NO_DEPENDENCIES}}
 // Microsoft Visual C++ generated include file.
-// Used by ManifestSchema.rc
+// Used by Schemas.rc
 
 // Next default values for new objects
 //
